+++ integration/service/oasis/stationfinder/cloudfinder/basic_finder_test.go
@@ -1,4 +1,4 @@
-package tnsearchfinder
+package cloudfinder
 
 import (
 	"reflect"
