+++ integration/service/oasis/stationfinder/cloudfinder/dest_station_finder_test.go
@@ -1,4 +1,4 @@
-package tnsearchfinder
+package cloudfinder
 
 import (
 	"reflect"
