+++ integration/service/oasis/stationfinder/cloudfinder/low_energy_location_station_finder.go
@@ -1,4 +1,4 @@
-package tnsearchfinder
+package cloudfinder
 
 import (
 	"github.com/Telenav/osrm-backend/integration/pkg/api/nav"
