+++ integration/service/oasis/stationfinder/cloudfinder/low_energy_location_station_finder_test.go
@@ -1,4 +1,4 @@
-package tnsearchfinder
+package cloudfinder
 
 import (
 	"reflect"
